requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.9.0
gunicorn==20.1.0
mysql-connector-python==8.0.32
Pillow==9.4.0
platformdirs==2.6.2
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.17.1
